Replace static occlusion namespaces with an owned injectable resolver
Review findings (Codex/Greptile/CodeRabbit converged): the caseless-enum
static namespaces violate the static-as-namespace ruling. One constructable
@MainActor SidebarResizerOcclusionResolver is now owned by ContentView with
the NSWindow.windowNumber(at:) lookup injected, so the full activation
decision path is testable. The unit suite is converted to Swift Testing and
covers the seam, including that drag and out-of-band paths never query the
window server.
